application_mila_assit/deploy-nas/backend/requirements.txt
# API Web
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
jinja2==3.1.6  # Templating HTML pour rapports d'évaluation

# Base de données
mysql-connector-python==9.1.0
SQLAlchemy==2.0.36

# Sécurité & Authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Rate Limiting
slowapi==0.1.9

# Intelligence Artificielle
# SUPPRIMÉ (9 Déc 2025): Architecture 4 containers
# Container 2 (API) ne fait plus d'IA (délégué à Container 3)
# Anciennes dépendances: sentence-transformers, faiss-cpu, torch, llama-cpp-python
# Ces packages sont maintenant dans llm-service/requirements.txt

# Utilitaires
python-dotenv==1.0.1
requests==2.32.3
httpx==0.28.1

# Tests
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
locust==2.32.4

# Logging & Monitoring
loguru==0.7.3

# Validation & Typage
pydantic[email]==2.10.3
